requirements.txt
speechrecognition
openai
python-dotenv
pyttsx3
pyaudio
winshell
